taint/test/mochitest/test_local_storage.html

<!DOCTYPE HTML>
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <meta charset="utf-8">
    <title>Check LocalStorage</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script "text/javascript">

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted");
      }

      function startTest() {
	  localStorage.setItem("untainted", "untainted string");
	  // Even untainted strings should be tainted as localStorage getItem is a source
	  check_tainted(localStorage.getItem("untainted"));

	  localStorage.setItem("tainted", String.tainted("tainted string"));
	  check_tainted(localStorage.getItem("tainted"));

	  SimpleTest.finish();
      }

      SimpleTest.waitForExplicitFinish();
      
    </script>
  </head>

  <body onload="startTest();">

</html>
